refactor: relocate new yaml utils to fit existing gestalt
Currently `js-yaml` dependency is part of common-server, so this commit
aims to cohere with that existing structure.
