+++ packages/common-all/src/utils/index.ts
@@ -6,7 +6,6 @@ import minimatch from "minimatch";
 import path from "path";
 import querystring from "querystring";
 import semver from "semver";
-import YAML from "js-yaml";
 import {
   ok,
   Ok,
@@ -31,7 +30,7 @@ import {
 } from "../constants";
 import { DENDRON_CONFIG } from "../constants/configs/dendronConfig";
 import { DendronError, ErrorMessages, IDendronError } from "../error";
-import { DHookDict, NoteChangeEntry, NoteProps, AnyJson } from "../types";
+import { DHookDict, NoteChangeEntry, NoteProps } from "../types";
 import { GithubConfig } from "../types/configs/publishing/github";
 import {
   DendronPublishingConfig,
@@ -1343,37 +1342,4 @@ export function globMatch(patterns: string[] | string, fname: string): boolean {
     return minimatch(fname, patterns);
   }
   return _.some(patterns, (pattern) => minimatch(fname, pattern));
-}
-
-export class YamlUtils {
-  static load = fromThrowable(YAML.load, (error) => {
-    return new DendronError({
-      message:
-        error instanceof YAML.YAMLException
-          ? `${error.name}: ${error.message}`
-          : `YAMLException`,
-      severity: ERROR_SEVERITY.FATAL,
-      ...(error instanceof Error && { innerError: error }),
-    });
-  });
-  static dump = fromThrowable(YAML.dump, (error) => {
-    return new DendronError({
-      message:
-        error instanceof YAML.YAMLException
-          ? `${error.name}: ${error.message}`
-          : `YAMLException`,
-      severity: ERROR_SEVERITY.FATAL,
-      ...(error instanceof Error && { innerError: error }),
-    });
-  });
-
-  static fromStr(str: string, overwriteDuplicate?: boolean) {
-    return this.load(str, {
-      schema: YAML.JSON_SCHEMA,
-      json: overwriteDuplicate ?? false,
-    }) as Result<AnyJson, DendronError>;
-  }
-  static toStr(data: any) {
-    return this.dump(data, { indent: 4, schema: YAML.JSON_SCHEMA });
-  }
 }
