+++ packages/common-server/src/DConfig.ts
@@ -14,7 +14,6 @@ import {
   RespV3,
   RespWithOptError,
   DendronConfig,
-  YamlUtils,
 } from "@dendronhq/common-all";
 import fs from "fs-extra";
 import _ from "lodash";
@@ -23,6 +22,7 @@ import path from "path";
 import { BackupKeyEnum, BackupService } from "./backup";
 import { DConfigLegacy } from "./oneoff/ConfigCompat";
 import { readYAML, writeYAML, writeYAMLAsync, readToString } from "./files";
+import { YamlUtils } from "./yaml";
 
 export enum LocalConfigScope {
   WORKSPACE = "WORKSPACE",
